requirements.txt
black==23.1.0
mypy==1.0.1
python-telegram-bot==20.1
ruff==0.0.254
